requirements.txt
absl-py==2.1.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
astunparse==1.6.3
attrs==24.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==3.1.0
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
fastapi==0.115.6
flatbuffers==24.3.25
fonttools==4.55.3
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
iniconfig==2.0.0
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
lightgbm==4.5.0
llvmlite==0.43.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numba==0.60.0
numpy==1.26.4
nvidia-nccl-cu12==2.23.4
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
optuna==4.1.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
protobuf==5.29.2
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.9
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.3.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.27.1
setuptools==75.6.0
shap==0.46.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
telegram==0.0.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
wsproto==1.2.0
xgboost==1.7.6
